+++ activemodel/lib/active_model.rb
@@ -45,6 +45,7 @@ module ActiveModel
   autoload :Dirty
   autoload :EachValidator, "active_model/validator"
   autoload :ForbiddenAttributesProtection
+  autoload :Inspect
   autoload :Lint
   autoload :Model
   autoload :Name, "active_model/naming"