+++ activerecord/lib/active_record/attribute_methods.rb
@@ -346,28 +346,6 @@ module ActiveRecord
       @attributes.to_hash
     end
 
-    # Returns an <tt>#inspect</tt>-like string for the value of the
-    # attribute +attr_name+. String attributes are truncated up to 50
-    # characters. Other attributes return the value of <tt>#inspect</tt>
-    # without modification.
-    #
-    #   person = Person.create!(name: 'David Heinemeier Hansson ' * 3)
-    #
-    #   person.attribute_for_inspect(:name)
-    #   # => "\"David Heinemeier Hansson David Heinemeier Hansson ...\""
-    #
-    #   person.attribute_for_inspect(:created_at)
-    #   # => "\"2012-10-22 00:15:07.000000000 +0000\""
-    #
-    #   person.attribute_for_inspect(:tag_ids)
-    #   # => "[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]"
-    def attribute_for_inspect(attr_name)
-      attr_name = attr_name.to_s
-      attr_name = self.class.attribute_aliases[attr_name] || attr_name
-      value = _read_attribute(attr_name)
-      format_for_inspect(attr_name, value)
-    end
-
     # Returns +true+ if the specified +attribute+ has been set by the user or by a
     # database load and is neither +nil+ nor <tt>empty?</tt> (the latter only applies
     # to objects that respond to <tt>empty?</tt>, most notably Strings). Otherwise, +false+.
@@ -516,22 +494,6 @@ module ActiveRecord
         attribute_names.delete_if do |name|
           (pk_attribute?(name) && id.nil?) ||
             column_for_attribute(name).virtual?
-        end
-      end
-
-      def format_for_inspect(name, value)
-        if value.nil?
-          value.inspect
-        else
-          inspected_value = if value.is_a?(String) && value.length > 50
-            "#{value[0, 50]}...".inspect
-          elsif value.is_a?(Date) || value.is_a?(Time)
-            %("#{value.to_fs(:inspect)}")
-          else
-            value.inspect
-          end
-
-          inspection_filter.filter_param(name, inspected_value)
         end
       end